fix(KEN-739): Address PR review - bind the cache too [no-changelog]
`load` read the cached identity, then let `with_access` read the store for
itself. A sign-in landing in between put the previous account's identity in
hand under a credential that passed every check: the call went out under
the new one, so the answer and the installed credential agreed, while the
cache belonged to neither.

Three routes served it. A 304 handed the cached value straight back, and a
refused status or a malformed body served it as offline. The transport
route named alongside them was already covered, because with no answer to
bind, that path compares the installed credential against the one the read
began with.

`with_access` now reports the credential the call opened under as well as
the one it ended under. `load` refuses unless the cache, the call and the
answer are all the same sign-in. A rotation opens under what the read saw
and saves its replacement before the call goes out, so both comparisons
hold across it.

`cached` and the ETag drawn from it were the only sign-in-scoped state read
before the call. The endpoint is not, and the credential store and the
generation file each already discard another endpoint's.

The test file crossed the 800-line ceiling, so it splits at its seam:
`me_client/main.rs` keeps the account states, the offline ladder and the
cache mechanics, and `me_client/sign_in_binding.rs` takes every test that
changes a credential under a read, with the transports and stores that
place the change at an exact point.

No changelog entry: changelog.d/fixed/KEN-739.md already states this
outcome for consumers, and this is the same symptom by another route.
